requirements.txt
python-dotenv
azure-devops